refactor: move ConsoleMonitor into separate extension
