+++ extensions/common/console-monitor/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.MonitorExtension
@@ -12,4 +12,4 @@
 #
 #
 
-org.eclipse.edc.monitor.opentelemetry.OtelMonitorExtension
+org.eclipse.edc.monitor.console.ConsoleMonitorExtension
