+++ settings.gradle.kts
@@ -164,6 +164,7 @@ include(":extensions:common:http:lib:jersey-providers-lib")
 include(":extensions:common:configuration:configuration-filesystem")
 include(":extensions:common:events:events-cloud-http")
 include(":extensions:common:http")
+include(":extensions:common:console-monitor")
 include(":extensions:common:otel-monitor")
 include(":extensions:common:http:jersey-core")
 include(":extensions:common:http:jersey-micrometer")
